Cache the feature scan results
On a solution with about 6000 test scenarios, Pickles.exe takes about 16 seconds on my machine to process.
This is with the following options:
Pickles.exe -f xxx -o xxx --trfmt=nunit3 --enableComments=false -df=json -lr=xxx

Profiling shows that a large part of this time is spent scanning the NUnit results. More precisely, scanning by feature name is done repeatedly on the xml results.
In this commit, the time is reduced to 5 seconds by scanning the xml once and storing the results in a C# lookup, which is optimized for fast-access by key.

Unit tests pass.
The resulting json files are identical (except "GeneratedOn" datetime, obviously).