Improve performance of VS Test results file parsing
On a VS test result file containing many tests, the current code is pretty slow. This was observed when migrating a codebase from NUnit3 result files to trx.

Here, based on profiling, we improve the performance by caching the parsing of the results:
- a lookup of tests per feature
- a dictionary of test results per test id

On a sample file with about 1000 tests, the performance improves from 40.2 seconds to 2.5 seconds.

This was tested with:
```
.\Pickles.exe -f xxx -o xxx --trfmt=vstest --enableComments=false -df=json -lr=xxx.trx
```

The resulting `pickledFeatures.json` file is identical to the one before these changes (except the `GeneratedOn` field).